Handle files in port tree merge glob, not just directories
The glob('*/\*/*') in merge_port_trees() matches both files and
directories. The cp -lr command assumed everything was a directory
(trailing /), failing on plain files like security/distinfo.
Check with os.path.isdir() and use cp -lf for files.

+++ build/tools/build-ports.py
@@ -159,9 +159,14 @@ def merge_port_trees():
                 else:
                     sh('cp -l ${p} ${PORTS_OVERLAY}/${portpath}')
             else:
-                sh('rm -rf ${PORTS_OVERLAY}/${portpath}')
-                sh('mkdir -p ${PORTS_OVERLAY}/${portpath}')
-                sh('cp -lr ${p}/ ${PORTS_OVERLAY}/${portpath}')
+                if os.path.isdir(p):
+                    sh('rm -rf ${PORTS_OVERLAY}/${portpath}')
+                    sh('mkdir -p ${PORTS_OVERLAY}/${portpath}')
+                    sh('cp -lr ${p}/ ${PORTS_OVERLAY}/${portpath}')
+                else:
+                    portdir = os.path.dirname(portpath)
+                    sh('mkdir -p ${PORTS_OVERLAY}/${portdir}')
+                    sh('cp -lf ${p} ${PORTS_OVERLAY}/${portpath}')
 
         if os.path.exists(uids):
             sh('rm -f ${PORTS_OVERLAY}/UIDs')
